(Incoming|Outgoing)MmsConnection interfaces
