+++ src/org/thoughtcrime/securesms/mms/LegacyMmsConnection.java
@@ -18,6 +18,7 @@ package org.thoughtcrime.securesms.mms;
 
 import android.content.Context;
 import android.net.ConnectivityManager;
+import android.telephony.TelephonyManager;
 import android.text.TextUtils;
 import android.util.Log;
 
@@ -50,19 +51,19 @@ import java.net.URL;
 import java.util.LinkedList;
 import java.util.List;
 
-public abstract class MmsConnection {
+@SuppressWarnings("deprecation")
+public abstract class LegacyMmsConnection {
   private static final String TAG = "MmsCommunication";
 
   protected final Context context;
   protected final Apn     apn;
 
-  protected MmsConnection(Context context, Apn apn) {
+  protected LegacyMmsConnection(Context context, Apn apn) {
     this.context = context;
     this.apn     = apn;
   }
 
-  public static Apn getApn(Context context, String apnName) throws ApnUnavailableException {
-    Log.w(TAG, "Getting MMSC params for apn " + apnName);
+  public static Apn getApn(Context context) throws ApnUnavailableException {
 
     try {
       Optional<Apn> params = ApnDatabase.getInstance(context)
@@ -77,6 +78,10 @@ public abstract class MmsConnection {
     } catch (IOException ioe) {
       throw new ApnUnavailableException("ApnDatabase threw an IOException", ioe);
     }
+  }
+
+  protected boolean isCdmaDevice() {
+    return ((TelephonyManager)context.getSystemService(Context.TELEPHONY_SERVICE)).getPhoneType() == TelephonyManager.PHONE_TYPE_CDMA;
   }
 
   protected static boolean checkRouteToHost(Context context, String host, boolean usingMmsRadio)
